requirements.txt
requests
beautifulsoup4
ortools
Flask
gunicorn
pyyaml
